[IMP] split files to improve readibility
