+++ pricelist_item_generator/models/__init__.py
@@ -1,1 +1,4 @@
+from . import product_condition
+from . import item_template
+from . import generator
 from . import pricelist
